Changes to fix tests and deal with incompatibilities
Plan B for having widely used test superclasses - have maven build a test-jar that can be depended on in addition to the main jar.

The test-jar is non-modular - this may not continue working if we ever *do* want to use the java module system within tests.
But it involves less arbitrarily carving code up into separate projects to avoid circular dependencies when a project contains a
test base class and also wants to use that class in its own tests without putting it in the main jar.

mesakit-map-measurements and mesakit-map-geography now both publish test-jar artifacts (GeometryUnitTest extends MeasurementUnitTest
and is then used pervasively in other mesakit projects).

I have added the test-jar as a test dependency to all of the projects that need GeometryUnitTest, so they should work as-is once
things are buildable.

I'm flying a bit blind here.  `Speed` and `Acceleration` relied on the removed `duration.asSeconds()` and `duration.asHours()`.
The most obvious replacements are `duration.dividedBy(Duration.ONE_SECOND)` and `duration.dividedBy(Duration.ONE_HOUR)`, but I have
no idea if the original returned an int or long or rounded value, making that not the right choice.

`Speed` also relied on the ability to get a milliseconds value out of a duration is gone.  I have added a field ONE_MILLISECOND to
`Duration` in KivaKit and used the same `dividedBy` logic to have a way of getting that here.  No idea if there is a less baroque way and
I'm just not seeing it.

This one I'm not going to try to solve, since there are many interfaces involved and it is not obvious which way of solving it is
the right one.  TLDR:

  * `Angle` appears to need to implement `QuantumComparable<Angle>` to get back its `isLessThan()`/`isGreaterThan()` methods.  I added that in.
  * It also appears that it _used to_ (?) implement Comparable<Angle> (an `@Override compareTo(Angle)` method is no longer inherited from anywhere). I added that in.
  * The result of that breaks `Latitude` and `Longitude`, which have a field `Range<Angle>`, and `Range` requires its type parameter to be `Comparable<Countable>`, not `Countable<Angle>`

Changing Range's signature to want its type parameter to extend `Comparable<Value>` or `Comparable<? super Value>` is not an option - it causes
`isLessThan()` and `isGreaterThan()` in QuantumComparable and Quantizable to become a name-clash - it seems like we are in a cul-de-sac where
`Value` must not be comparable on itself, but (judging from the implementation of `Angle.compareTo()`), it has to be.

Since you know the intent of these interfaces far better than I do, this type-system Rubik's cube I leave to you.

But the test stuff is fixed, and so whenever this can be untangled, it should all _just work_ (famous last words).